libs/config/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neurolab-config"
version = "0.1.0"
description = "Cluster environment configuration for local, Expanse, and Delta HPC systems"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["neurolab.*"]

[tool.setuptools.package-data]
"neurolab.config" = ["profiles/*.yaml"]
